Added dynamic schema generation and type file
Wrote a script that grabs the schema defined by jsonresume.org and saves
it locally as schema.json. In addition got a library to dynamically
generate a type file for the schema from the local schema.json then
added them to global scope. This will allow for better scalability
ideally.